+++ src/pics.rs
@@ -5,6 +5,7 @@ use crate::concurrent_map::ConcurrentMap;
 use std::process::Command;
 use crate::args::AppArgs;
 use crate::Progs;
+use crate::formats::Format;
 
 pub type PicsMap = ConcurrentMap<String, Option<Vec<u8>>>;
 
@@ -12,35 +13,43 @@ pub fn ffmpeg_conv_pic_args(app_args: &AppArgs) -> Vec<String> {
     let fmt = format!(
         "scale='w=min({},iw)':h='min({},ih)':force_original_aspect_ratio=decrease:flags=lanczos",
         app_args.max_pic_width, app_args.max_pic_height);
-    let q = app_args.pic_quality.to_string();
-    return vec!["-vf".to_string(), fmt, "-qmin".to_string(), "1".to_string(), "-q:v".to_string(), q];
+    let mut conv_args = vec!["-vf".to_string(), fmt];
+    let quality_args = app_args.output_ext_type.pic_quality_args(app_args.pic_quality);
+    conv_args.extend(quality_args);
+    return conv_args;
 }
 
 fn conv_pic(pic_file: &str, app_args: &AppArgs, progs: &Progs) -> Option<Vec<u8>> {
-    let pic_args = ffmpeg_conv_pic_args(app_args);
-    let pic_args = pic_args.iter().map(String::as_str).collect::<Vec<&str>>();
+    match app_args.output_ext_type {
+        Format::OGG => return std::fs::read(pic_file).ok(),
 
-    let mut args = vec![
-        "-i", pic_file,
-        "-f", "mjpeg"
-    ];
+        Format::MP3 => {
+            let pic_args = ffmpeg_conv_pic_args(app_args);
+            let pic_args = pic_args.iter().map(String::as_str).collect::<Vec<&str>>();
 
-    args.extend(pic_args);
-    args.push("-");
+            let mut args = vec![
+                "-i", pic_file,
+                "-f", "mjpeg"
+            ];
 
-    let args_str = shell_words::join(&args);
-    println!("PIC {}: {} {}", pic_file, &progs.ffmpeg_bin, args_str);
+            args.extend(pic_args);
+            args.push("-");
 
-    if app_args.dry_run {
-        return Some(Vec::new());
+            let args_str = shell_words::join(&args);
+            println!("PIC {}: {} {}", pic_file, &progs.ffmpeg_bin, args_str);
+
+            if app_args.dry_run {
+                return Some(Vec::new());
+            }
+
+            let output = Command::new(&progs.ffmpeg_bin).args(args).output().ok()?;
+            if output.status.code()? != 0 {
+                println!("PIC {}: {}", pic_file, std::str::from_utf8(&output.stderr).unwrap());
+                return None;
+            }
+            return Some(output.stdout);
+        }
     }
-
-    let output = Command::new(&progs.ffmpeg_bin).args(args).output().ok()?;
-    if output.status.code()? != 0 {
-        println!("PIC {}: {}", pic_file, std::str::from_utf8(&output.stderr).unwrap());
-        return None;
-    }
-    return Some(output.stdout);
 }
 
 impl PicsMap {