+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # musbconv - CHANGELOG
 
 
+## unreleased
+
+- Fixed: cannot convert to OGG with an external cover art
+- Removed: support for any output format except MP3 and OGG
+
+
 ## v0.2.0 (May 16, 2021)
 
 - Fixed: ffmpeg errors are not captured when a cover art is piped